refactor: split multiple React components of TabDetails into separate files
