+++ src/pages/TabDetails/CycleList/index.ts
@@ -1,0 +1,3 @@
+import CycleList from "./CycleList";
+
+export default CycleList;
