+++ src/pages/TabDetails/TabDetails.css
