+++ src/pages/TabDetails/index.ts
@@ -1,0 +1,3 @@
+import TabDetails from "./TabDetails";
+
+export default TabDetails;
