+++ src/pages/TabDetails/styles.ts
@@ -1,0 +1,16 @@
+export const lenCycleStyle = {
+  fontSize: "13px",
+  color: "var(--ion-color-black)",
+  textAlign: "left",
+} as const;
+
+export const datesStyle = {
+  fontSize: "11px",
+  color: "var(--ion-color-medium)",
+  textAlign: "left",
+} as const;
+
+export const progressBarStyle = {
+  marginTop: "5px",
+  marginBottom: "5px",
+} as const;
