+++ src/utils/progress-bar.ts
@@ -1,0 +1,3 @@
+export function getNormalizedProgress(value: number, maxLength: number) {
+  return (value / maxLength) * 0.95;
+}
